Add api , item, cart and item cart

+++ cart-order-api/src/main/java/com/tech/repository/UserRepository.java
@@ -4,7 +4,7 @@ import com.tech.domain.User;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface UserRepository extends MongoRepository<User, Long> {
+public interface UserRepository extends MongoRepository<User, Integer> {
 
     User findUserByEmail(String email);
